Feature updates and Docker work
