+++ .gitignore
@@ -1,9 +1,2 @@
-.project
-.settings/
-.vscode/
-.idea/
-qaframework.iml
-target/
-.classpath
-.vscode
-.factorypath
+/target/
+.idea
