+++ src/test/resources/ci.test.properties
@@ -1,0 +1,5 @@
+webapp.url=http://isanteplus:8080/openmrs
+login.username=admin
+login.password=Admin123
+webdriver=chrome
+headless=true
