+++ src/test/resources/test.properties
@@ -1,4 +1,4 @@
-webapp.url=http://isanteplusdemo.com:8080/openmrs
+webapp.url=http://isanteplus.sedish-haiti.org/openmrs
 login.username=admin
 login.password=Admin123
 webdriver=chrome
